UserControl/requirements.txt
pyserial
pyyaml
pandas
matplotlib
numpy
scipy
jsonschema
